app/requirements.txt
amqp==2.6.0
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==1.2.0
beautifulsoup4==4.8.1
billiard==3.6.3.0
bs4==0.0.1
boto
celery==4.4.6
celery-progress==0.0.14
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
chromedriver-binary==86.0.4240.22
cryptography==2.8
cssselect==1.1.0
Django==2.2.5
django-appconf==1.0.3
django-axes==5.4.1
django-bootstrap4==1.1.1
django-celery-results==1.2.1
django-ipware==3.0.0
django-redis==4.12.1
django-sass-processor==0.7.5
django-compressor==2.3
django-ses
# dnspython==2.0.0
dnspython==1.15.0

# eventlet==0.25.2
eventlet

fake-useragent==0.1.11
flower==0.9.5
future==0.18.2

# gevent==20.9.0
gevent

googlemaps==4.3.1
# greenlet==0.4.16
greenlet

gunicorn == 20.0.4
humanize==2.5.0
idna==2.8
importlib-metadata==1.7.0
kombu==4.6.11
libsass==0.19.4
lxml==4.5.0
# mkl-fft==1.0.15
# mkl-random==1.1.0
# mkl-service==2.3.0
monotonic==1.5
# numpy==1.14.3
#-e git+https://github.com/IntelPython/mkl_fft.git#egg=mkl_fft
#-e git+https://github.com/IntelPython/mkl_random.git#egg=mkl_random
#-e git+https://github.com/IntelPython/mkl_service.git#egg=mkl_service
olefile==0.46
# pandas
parse==1.15.0
pillow==7.2.0
plotly==4.5.0
prometheus-client==0.8.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
pyee==7.0.1
pygeocoder==1.2.5
pyOpenSSL==19.1.0
pyppeteer==0.0.25
pyquery==1.4.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.3
requests==2.22.0
requests-html==0.10.0
retrying==1.3.3
rjsmin==1.1.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.19
sqlparse==0.3.0
tornado==6.0.4
tqdm==4.43.0
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
w3lib==1.21.0
websockets==8.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
